Relocate `_DefaultStack` to a new module.
Removes the `_DefaultStack` class from `ops.py`
and renames it to `DefaultStack`. Additionally relocates
the `default_session()` and `get_default_session()` methods
to this new module.

These methods are widely used outside of `ops` and in
some cases are a module's only dependency on `ops`.
It is also necessary to move these methods in preparation
for further refactors of `ops.py`.

To avoid immediate disruptive changes in the wider TensorFlow
codebase, `get_default_session` is aliased in `ops.py`.

PiperOrigin-RevId: 514825999