Add `-mac_excluded` to TF build/test tag filters
Currently, `no_mac` or `nomac` is used to exclude a test from running in the mac environment. However, it is difficult to distinguish between temporary and permanent exclusions. For example, a test may be disabled temporarily if it is broken, or it may be designed to not run on mac environment permanently. To address this issue, we introduce a new tag `mac_excluded` for platform exclusion design. `no_mac` or `nomac` will now be considered to disable broken tests, while `mac_excluded` will be used to permanently exclude a test from running on mac environment.

PiperOrigin-RevId: 515676485