Improve side effect modeling of XLA launch op.
There are currently no side effect traits associated with tf.XlaLaunch and as a result its side effect is treated as unknown, creating unwanted dependency that may result in deadlocks when it causes transitive dependency between collective ops that should execute in parallel. This implements MemoryEffectsOpInterface to relax the conservatism.

PiperOrigin-RevId: 516624982