[TFRT] Make eager context register-able.
Main motivation is to make TFRT eager runtime registration based so as to shrink binary size (TFRT eager runtime is not registered by default).

Note: c_api_tfrt is not removed from tensorflow/c/eager:c_api in this change yet, due to some tagging issue. This is to be cleaned up in follow up changes.
PiperOrigin-RevId: 516679347