Remove obsolete CHECK lines.
There used to be a time where fusion instructions would store all called
computations that are called from within their fusion computation. So we didn't
allow to update any called computation inside a fusion computation, because it
would have required to update the called_computations list of the fusion
instruction (note, that the comment said called_instructions, but that was
wrong all along). However now called_computations of a fusion instruction just
contains the fusion computation, so we don't need the CHECK lines anymore.
Instead, adjust the check that called_computations for fusion instructions are non-empty to check for size 1.

PiperOrigin-RevId: 516784255